wado-compiler: extract type helpers from cm_binding into a submodule
Begins the cm_binding.rs refactor by carving out type-level utilities
(WASI -> TypeId, CM ABI flatten / size / align, struct/variant flat
helpers) into synthesis/cm_binding/types.rs. The parent file shrinks
from 10,117 to 9,156 lines while the public API is preserved through
re-exports.
